+++ src/maestral/client.py
@@ -10,8 +10,7 @@ import os
 import re
 import time
 import functools
-import contextlib
-import threading
+from contextlib import contextmanager, closing
 from datetime import datetime, timezone
 from typing import (
     Callable,
@@ -32,6 +31,12 @@ from dropbox import files, sharing, users, common
 from dropbox import Dropbox, create_session, exceptions
 from dropbox.oauth import DropboxOAuth2FlowNoRedirect
 from dropbox.session import API_HOST
+from dropbox.dropbox_client import (
+    BadInputException,
+    RouteResult,
+    RouteErrorResult,
+    USER_AUTH,
+)
 
 # local imports
 from . import __version__
@@ -89,14 +94,82 @@ PRT = TypeVar("PRT", ListFolderResult, ListSharedLinkResult)
 P = ParamSpec("P")
 T = TypeVar("T")
 
-major_minor_version = ".".join(__version__.split(".")[:2])
-USER_AGENT = f"Maestral/v{major_minor_version}"
+USER_AGENT = f"Maestral/v{__version__}"
 
 
 def get_hash(data: bytes) -> str:
     hasher = DropboxContentHasher()
     hasher.update(data)
     return hasher.hexdigest()
+
+
+class _DropboxSDK(Dropbox):
+    def request_json_string(
+        self,
+        host: str,
+        func_name: str,
+        route_style: str,
+        request_json_arg: bytes,
+        auth_type: str,
+        request_binary: bytes | Iterator[bytes] | None,
+        timeout: float | None = None,
+    ) -> RouteResult | RouteErrorResult:
+        # Custom handling to allow for streamed and chunked uploads. This is mostly
+        # reproduced from the parent function but without limiting the request body
+        # to bytes only.
+        if route_style == self._ROUTE_STYLE_UPLOAD:
+            fq_hostname = self._host_map[host]
+            url = self._get_route_url(fq_hostname, func_name)
+
+            auth_types = auth_type.replace(" ", "").split(",")
+
+            if USER_AUTH not in auth_types:
+                raise BadInputException("Unhandled auth type: {}".format(auth_type))
+
+            headers = {
+                "User-Agent": self._user_agent,
+                "Authorization": f"Bearer {self._oauth2_access_token}",
+            }
+
+            if self._headers:
+                headers.update(self._headers)
+
+            headers["Content-Type"] = "application/octet-stream"
+            headers["Dropbox-API-Arg"] = request_json_arg
+            body = request_binary
+
+            if timeout is None:
+                timeout = self._timeout
+
+            r = self._session.post(
+                url,
+                headers=headers,
+                data=body,
+                stream=False,
+                verify=True,
+                timeout=timeout,
+            )
+
+            self.raise_dropbox_error_for_resp(r)
+
+            if r.status_code in (403, 404, 409):
+                raw_resp = r.content.decode("utf-8")
+                request_id = r.headers.get("x-dropbox-request-id")
+                return RouteErrorResult(request_id, raw_resp)
+
+            raw_resp = r.content.decode("utf-8")
+            return RouteResult(raw_resp)
+
+        else:
+            return super().request_json_string(
+                host,
+                func_name,
+                route_style,
+                request_json_arg,
+                auth_type,
+                request_binary,
+                timeout,
+            )
 
 
 class DropboxClient:
@@ -124,6 +197,10 @@ class DropboxClient:
     :param timeout: Timeout for individual requests. Defaults to 100 sec if not given.
     :param session: Optional requests session to use. If not given, a new session will
         be created with :func:`dropbox.dropbox_client.create_session`.
+    :param bandwidth_limit_up: Maximum bandwidth to use for uploads in bytes/sec. Will
+        be enforced over all concurrent uploads.
+    :param bandwidth_limit_down: Maximum bandwidth to use for downloads in bytes/sec.
+        Will be enforced over all concurrent downloads.
     """
 
     SDK_VERSION: str = "2.0"
@@ -131,7 +208,9 @@ class DropboxClient:
     MAX_TRANSFER_RETRIES = 3
     MAX_LIST_FOLDER_RETRIES = 3
 
-    _dbx: Dropbox | None
+    UPLOAD_REQUEST_CHUNK_SIZE = 4194304
+
+    _dbx: _DropboxSDK | None
 
     def __init__(
         self,
@@ -139,6 +218,8 @@ class DropboxClient:
         cred_storage: CredentialStorage,
         timeout: float = 100,
         session: requests.Session | None = None,
+        bandwidth_limit_up: float = 0,
+        bandwidth_limit_down: float = 0,
     ) -> None:
         self.config_name = config_name
         self._auth_flow: DropboxOAuth2FlowNoRedirect | None = None
@@ -152,12 +233,77 @@ class DropboxClient:
         self._timeout = timeout
         self._session = session or create_session()
         self._backoff_until = 0
-        self._dbx: Dropbox | None = None
-        self._dbx_base: Dropbox | None = None
+        self._dbx: _DropboxSDK | None = None
+        self._dbx_base: _DropboxSDK | None = None
         self._cached_account_info: FullAccount | None = None
         self._namespace_id = self._state.get("account", "path_root_nsid")
         self._is_team_space = self._state.get("account", "path_root_type") == "team"
-        self._lock = threading.Lock()
+
+        # Throttling infra
+        self.bandwidth_limit_up = bandwidth_limit_up
+        self.bandwidth_limit_down = bandwidth_limit_down
+
+        # The Dropbox SDK currently only support streamed downloads but not uploads.
+        # This means that we have to supply binary data in chunks and pause in between.
+        # However, for efficiency and API call limits, the chunk size should not be too
+        # small. Choose 4 Mb as a compromise. This means that uploads throttled to
+        # < 4 Mb/sec will appear choppy instead of smooth.
+        # The alternative approach of limiting at the `socket.send` step is too
+        # cumbersome given the abstraction layers of urllib3, requests, and Dropbox SDK.
+        self.download_chunk_size = 2048  # 2 Kb
+        self.upload_chunk_size = 2048  # 4 Mb
+
+        self._num_downloads = 0
+        self._num_uploads = 0
+
+    @contextmanager
+    def _register_download(self) -> Iterator[None]:
+        self._num_downloads += 1
+        try:
+            yield
+        finally:
+            self._num_downloads -= 1
+
+    @contextmanager
+    def _register_upload(self) -> Iterator[None]:
+        self._num_uploads += 1
+        try:
+            yield
+        finally:
+            self._num_uploads -= 1
+
+    def _throttled_download_iter(self, iterator: Iterator[T]) -> Iterator[T]:
+        for i in iterator:
+            if self.bandwidth_limit_down == 0:
+                yield i
+            else:
+                tick = time.monotonic()
+                yield i
+                tock = time.monotonic()
+
+                speed_per_download = self.bandwidth_limit_down / self._num_downloads
+                target_tock = tick + self.download_chunk_size / speed_per_download
+
+                wait_time = target_tock - tock
+                if wait_time > 0.00005:  # don't sleep for < 50 ns
+                    time.sleep(wait_time)
+
+    def _throttled_upload_iter(self, data: bytes) -> Iterator[bytes] | bytes:
+        pos = 0
+        while pos < len(data):
+            tick = time.monotonic()
+            yield data[pos : pos + self.upload_chunk_size]
+            tock = time.monotonic()
+
+            pos += self.upload_chunk_size
+
+            if self.bandwidth_limit_up > 0:
+                speed_per_upload = self.bandwidth_limit_up / self._num_uploads
+                target_tock = tick + self.upload_chunk_size / speed_per_upload
+
+                wait_time = target_tock - tock
+                if wait_time > 0.00005:  # don't sleep for < 50 ns
+                    time.sleep(wait_time)
 
     def _retry_on_error(  # type: ignore
         error_cls: type[Exception],
@@ -220,18 +366,18 @@ class DropboxClient:
     # ---- Linking API -----------------------------------------------------------------
 
     @property
-    def dbx_base(self) -> Dropbox:
+    def dbx_base(self) -> _DropboxSDK:
         """The underlying Dropbox SDK instance without namespace headers."""
         if not self._dbx_base:
             self._init_sdk()
-        return self._dbx_base
+        return cast(_DropboxSDK, self._dbx_base)
 
     @property
-    def dbx(self) -> Dropbox:
+    def dbx(self) -> _DropboxSDK:
         """The underlying Dropbox SDK instance with namespace headers."""
         if not self._dbx:
             self._init_sdk()
-        return self._dbx
+        return cast(_DropboxSDK, self._dbx)
 
     @property
     def linked(self) -> bool:
@@ -341,45 +487,44 @@ class DropboxClient:
         :raises RuntimeError: if token is not available from storage and no token is
             passed as an argument.
         """
-        with self._lock:
-            if not (token or self._cred_storage.token):
-                raise NotLinkedError(
-                    "No auth token set", "Please link a Dropbox account first."
-                )
+        if not (token or self._cred_storage.token):
+            raise NotLinkedError(
+                "No auth token set", "Please link a Dropbox account first."
+            )
 
-            token = token or self._cred_storage.token
-            token_type = token_type or self._cred_storage.token_type
+        token = token or self._cred_storage.token
+        token_type = token_type or self._cred_storage.token_type
 
-            if token_type is TokenType.Offline:
-                # Initialise Dropbox SDK.
-                self._dbx_base = Dropbox(
-                    oauth2_refresh_token=token,
-                    app_key=DROPBOX_APP_KEY,
-                    session=self._session,
-                    user_agent=USER_AGENT,
-                    timeout=self._timeout,
-                )
-            else:
-                # Initialise Dropbox SDK.
-                self._dbx_base = Dropbox(
-                    oauth2_access_token=token,
-                    app_key=DROPBOX_APP_KEY,
-                    session=self._session,
-                    user_agent=USER_AGENT,
-                    timeout=self._timeout,
-                )
+        if token_type is TokenType.Offline:
+            # Initialise Dropbox SDK.
+            self._dbx_base = _DropboxSDK(
+                oauth2_refresh_token=token,
+                app_key=DROPBOX_APP_KEY,
+                session=self._session,
+                user_agent=USER_AGENT,
+                timeout=self._timeout,
+            )
+        else:
+            # Initialise Dropbox SDK.
+            self._dbx_base = _DropboxSDK(
+                oauth2_access_token=token,
+                app_key=DROPBOX_APP_KEY,
+                session=self._session,
+                user_agent=USER_AGENT,
+                timeout=self._timeout,
+            )
 
-            # If namespace_id was given, use the corresponding namespace, otherwise
-            # default to the home namespace.
-            if self._namespace_id:
-                root_path = common.PathRoot.root(self._namespace_id)
-                self._dbx = self._dbx_base.with_path_root(root_path)
-            else:
-                self._dbx = self._dbx_base
+        # If namespace_id was given, use the corresponding namespace, otherwise
+        # default to the home namespace.
+        if self._namespace_id:
+            root_path = common.PathRoot.root(self._namespace_id)
+            self._dbx = self._dbx_base.with_path_root(root_path)
+        else:
+            self._dbx = self._dbx_base
 
-            # Set our own logger for the Dropbox SDK.
-            self._dbx._logger = self._dropbox_sdk_logger
-            self._dbx_base._logger = self._dropbox_sdk_logger
+        # Set our own logger for the Dropbox SDK.
+        self.dbx._logger = self._dropbox_sdk_logger
+        self.dbx_base._logger = self._dropbox_sdk_logger
 
     @property
     def account_info(self) -> FullAccount:
@@ -446,7 +591,7 @@ class DropboxClient:
 
         path_root = common.PathRoot.root(root_nsid)
         self._dbx = self.dbx_base.with_path_root(path_root)
-        self._dbx._logger = self._dropbox_sdk_logger
+        self.dbx._logger = self._dropbox_sdk_logger
 
         if isinstance(root_info, UserRootInfo):
             actual_root_type = "user"
@@ -620,19 +765,21 @@ class DropboxClient:
         :returns: Metadata of downloaded item.
         :raises DataCorruptionError: if data is corrupted during download.
         """
-        chunk_size = 2**13
 
         with convert_api_errors(dbx_path=dbx_path):
             md, http_resp = self.dbx.files_download(dbx_path)
 
-            with contextlib.closing(http_resp):
+            with closing(http_resp):
                 with open(local_path, "wb", opener=opener_no_symlink) as f:
                     hasher = DropboxContentHasher()
                     wrapped_f = StreamHasher(f, hasher)
-                    for c in http_resp.iter_content(chunk_size):
-                        wrapped_f.write(c)
-                        if sync_event:
-                            sync_event.completed = wrapped_f.tell()
+                    with self._register_download():
+                        for c in self._throttled_download_iter(
+                            http_resp.iter_content(self.download_chunk_size)
+                        ):
+                            wrapped_f.write(c)
+                            if sync_event:
+                                sync_event.completed = wrapped_f.tell()
 
                     local_hash = hasher.hexdigest()
 
@@ -657,7 +804,6 @@ class DropboxClient:
         self,
         local_path: str,
         dbx_path: str,
-        chunk_size: int = 5 * 10**6,
         write_mode: WriteMode = WriteMode.Add,
         update_rev: str | None = None,
         autorename: bool = False,
@@ -668,8 +814,6 @@ class DropboxClient:
 
         :param local_path: Path of local file to upload.
         :param dbx_path: Path to save file on Dropbox.
-        :param chunk_size: Maximum size for individual uploads. If larger than 150 MB,
-            it will be set to 150 MB.
         :param write_mode: Your intent when writing a file to some path. This is used to
             determine what constitutes a conflict and what the autorename strategy is.
             This is used to determine what
@@ -704,8 +848,6 @@ class DropboxClient:
         :returns: Metadata of uploaded file.
         :raises DataCorruptionError: if data is corrupted during upload.
         """
-        chunk_size = clamp(chunk_size, 10**5, 150 * 10**6)
-
         if write_mode is WriteMode.Add:
             dbx_write_mode = files.WriteMode.add
         elif write_mode is WriteMode.Overwrite:
@@ -723,46 +865,43 @@ class DropboxClient:
             # Dropbox SDK takes naive datetime in UTC
             mtime_dt = datetime.utcfromtimestamp(stat.st_mtime)
 
-            if stat.st_size <= chunk_size:
-                # Upload all at once.
-
-                res = self._upload_helper(
-                    local_path,
-                    dbx_path,
-                    mtime_dt,
-                    dbx_write_mode,
-                    autorename,
-                    sync_event,
-                )
-
-            else:
-                # Upload in chunks.
-                # Note: We currently do not support resuming interrupted uploads.
-                # Dropbox keeps upload sessions open for 48h so this could be done in
-                # the future.
-
-                with open(local_path, "rb", opener=opener_no_symlink) as f:
-                    session_id = self._upload_session_start_helper(
-                        f, chunk_size, dbx_path, sync_event
-                    )
-
-                    while stat.st_size - f.tell() > chunk_size:
-                        self._upload_session_append_helper(
-                            f, session_id, chunk_size, dbx_path, sync_event
-                        )
-
-                    res = self._upload_session_finish_helper(
-                        f,
-                        session_id,
-                        chunk_size,
-                        # Commit info.
+            with self._register_upload():
+                if stat.st_size <= self.UPLOAD_REQUEST_CHUNK_SIZE:
+                    # Upload all at once.
+                    res = self._upload_helper(
+                        local_path,
                         dbx_path,
                         mtime_dt,
                         dbx_write_mode,
                         autorename,
-                        # Commit info end.
                         sync_event,
                     )
+                else:
+                    # Upload in chunks.
+                    # Note: We currently do not support resuming interrupted uploads.
+                    # Dropbox keeps upload sessions open for 48h so this could be done
+                    # in the future.
+                    with open(local_path, "rb", opener=opener_no_symlink) as f:
+                        session_id = self._upload_session_start_helper(
+                            f, dbx_path, sync_event
+                        )
+
+                        while stat.st_size - f.tell() > self.UPLOAD_REQUEST_CHUNK_SIZE:
+                            self._upload_session_append_helper(
+                                f, session_id, dbx_path, sync_event
+                            )
+
+                        res = self._upload_session_finish_helper(
+                            f,
+                            session_id,
+                            # Commit info.
+                            dbx_path,
+                            mtime_dt,
+                            dbx_write_mode,
+                            autorename,
+                            # Commit info end.
+                            sync_event,
+                        )
 
         return convert_metadata(res)
 
@@ -781,7 +920,7 @@ class DropboxClient:
 
             with convert_api_errors(dbx_path=dbx_path, local_path=local_path):
                 md = self.dbx.files_upload(
-                    data,
+                    self._throttled_upload_iter(data),
                     dbx_path,
                     client_modified=client_modified,
                     content_hash=get_hash(data),
@@ -798,17 +937,16 @@ class DropboxClient:
     def _upload_session_start_helper(
         self,
         f: BinaryIO,
-        chunk_size: int,
         dbx_path: str,
         sync_event: SyncEvent | None,
     ) -> str:
         initial_offset = f.tell()
-        data = f.read(chunk_size)
+        data = f.read(self.UPLOAD_REQUEST_CHUNK_SIZE)
 
         try:
             with convert_api_errors(dbx_path=dbx_path):
                 session_start = self.dbx.files_upload_session_start(
-                    data, content_hash=get_hash(data)
+                    self._throttled_upload_iter(data), content_hash=get_hash(data)
                 )
         except Exception:
             # Return to previous position in file.
@@ -825,12 +963,11 @@ class DropboxClient:
         self,
         f: BinaryIO,
         session_id: str,
-        chunk_size: int,
         dbx_path: str,
         sync_event: SyncEvent | None,
     ) -> None:
         initial_offset = f.tell()
-        data = f.read(chunk_size)
+        data = f.read(self.UPLOAD_REQUEST_CHUNK_SIZE)
 
         cursor = files.UploadSessionCursor(
             session_id=session_id,
@@ -840,7 +977,9 @@ class DropboxClient:
         try:
             with convert_api_errors(dbx_path=dbx_path):
                 self.dbx.files_upload_session_append_v2(
-                    data, cursor, content_hash=get_hash(data)
+                    self._throttled_upload_iter(data),
+                    cursor,
+                    content_hash=get_hash(data),
                 )
         except exceptions.DropboxException as exc:
             error = getattr(exc, "error", None)
@@ -865,7 +1004,6 @@ class DropboxClient:
         self,
         f: BinaryIO,
         session_id: str,
-        chunk_size: int,
         dbx_path: str,
         client_modified: datetime,
         mode: files.WriteMode,
@@ -873,13 +1011,9 @@ class DropboxClient:
         sync_event: SyncEvent | None,
     ) -> files.FileMetadata:
         initial_offset = f.tell()
-        data = f.read(chunk_size)
-
-        if len(data) > chunk_size:
-            raise RuntimeError("Too much data left to finish the session")
+        data = f.read(self.UPLOAD_REQUEST_CHUNK_SIZE)
 
         # Finish upload session and return metadata.
-
         cursor = files.UploadSessionCursor(
             session_id=session_id,
             offset=initial_offset,
@@ -894,7 +1028,10 @@ class DropboxClient:
         try:
             with convert_api_errors(dbx_path=dbx_path):
                 md = self.dbx.files_upload_session_finish(
-                    data, cursor, commit, content_hash=get_hash(data)
+                    self._throttled_upload_iter(data),
+                    cursor,
+                    commit,
+                    content_hash=get_hash(data),
                 )
         except exceptions.DropboxException as exc:
             error = getattr(exc, "error", None)
@@ -1039,7 +1176,6 @@ class DropboxClient:
             conflict.
         :returns: Metadata of created folder.
         """
-
         with convert_api_errors(dbx_path=dbx_path):
             res = self.dbx.files_create_folder_v2(dbx_path, autorename)